pubsub/requirements.txt

google-cloud-pubsub
google-cloud-functions
google-cloud-storage
avro
protobuf
